requirements.txt
flask
flask-sqlalchemy
flask-cors
requests
python-dotenv
psycopg2-binary
gunicorn
